Revert "BOTLIB: fixed friction in AAS_ClientMovementPrediction"
This reverts commit dec81dce6a9318bdd5a38ad5534897ed2ddf2657.

KuehnhammerTobias:
Just a NOTE:
As long as you use seperate botlib files as a dependency for your BSPC tool,
this fix shouldn't be applied! It destroys the ability to compute jump reachabilities.
The issue in short: the bot predicts it's jump route using 'AAS_ClientMovementPrediction'
(onground is set to qtrue), this means he uses 'phys_friction' (6) to predict the jump.
Though, actually he is NOT on ground when he is jumping, but he is also not swimming this
means jumping through air needs no (or at least nearly zero) friction, this seems to be
the reason they choose 'phys_waterfriction' (1) for when not swimming (assuming he is
jumping). All this becomes even more complicated because the bot can be onground AND
swimming (walking under water).
Additionally 'onground' is abused inside 'AAS_ClientMovementPrediction' as it seems.
Anyways, the opposite is true if the bot uses 'AAS_ClientMovementPrediction' for
non-precomputed actions (dynamically). For example in 'BotWalkInDirection', caused
if a bot wants to avoid obstacles or during attack moves. In this case
'AAS_ClientMovementPrediction' is called (from botlib) to check for gaps, hazards,
and whatnot. Here it makes sense to use the correct 'code' since the jump reachabilities
are already (pre)computed. All this is hacky, a real fix is on the run, though i'm lazy...

Here is a showcase of the problem, and where/how to fix:
[bspc-botlib-friction-fix-showcase.txt](https://github.com/PadWorld-Entertainment/bspc/files/6605536/bspc-botlib-friction-fix-showcase.txt)
Rename this file to .patch or .diff, but don't apply it to WoP, it's just for demonstration purposes.